+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/ColorScreen.java
@@ -7,7 +7,7 @@ import club.iananderson.seasonhud.client.gui.components.sliders.rgb.BlueSlider;
 import club.iananderson.seasonhud.client.gui.components.sliders.rgb.GreenSlider;
 import club.iananderson.seasonhud.client.gui.components.sliders.rgb.RedSlider;
 import club.iananderson.seasonhud.client.gui.components.sliders.rgb.RgbSlider;
-import club.iananderson.seasonhud.config.Config;
+import club.iananderson.seasonhud.config.SeasonHudClient;
 import club.iananderson.seasonhud.impl.seasons.Seasons;
 import club.iananderson.seasonhud.util.Rgb;
 import java.util.ArrayList;
@@ -22,7 +22,7 @@ import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
 
 public class ColorScreen extends SeasonHudScreen {
-  private static final Component SCREEN_TITLE = Component.translatable("menu.seasonhud.color.title");
+  private static final Component SCREEN_TITLE = Common.translatedText("menu.seasonhud.color.title");
   private final List<ColorEditBox> seasonBoxes = new ArrayList<>();
   private int x;
   private int y;
@@ -40,8 +40,9 @@ public class ColorScreen extends SeasonHudScreen {
 
   private static EnumSet<Seasons> seasonListSet() {
     EnumSet<Seasons> set = Seasons.SEASONS_ENUM_LIST.clone();
+    set.remove(Seasons.NULL);
 
-    if (!Config.getShowTropicalSeason() || Common.fabricSeasonsLoaded()) {
+    if (!SeasonHudClient.getShowTropicalSeason() || Common.fabricSeasonsLoaded()) {
       set.remove(Seasons.DRY);
       set.remove(Seasons.WET);
     }
@@ -50,7 +51,7 @@ public class ColorScreen extends SeasonHudScreen {
   }
 
   public void loadConfig() {
-    seasonColor = Config.getEnableSeasonNameColor();
+    seasonColor = SeasonHudClient.getEnableSeasonNameColor();
   }
 
   @Override
@@ -66,7 +67,7 @@ public class ColorScreen extends SeasonHudScreen {
 
   @Override
   public void onClose() {
-    Config.setEnableSeasonNameColor(seasonColor);
+    SeasonHudClient.setEnableSeasonNameColor(seasonColor);
     super.onClose();
   }
 
@@ -152,11 +153,11 @@ public class ColorScreen extends SeasonHudScreen {
     });
 
     //Buttons
-    CycleButton<Boolean> seasonColorButton = CycleButton.onOffBuilder(Config.getEnableSeasonNameColor())
-        .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.color.enableSeasonNameColor.tooltip")))
+    CycleButton<Boolean> seasonColorButton = CycleButton.onOffBuilder(SeasonHudClient.getEnableSeasonNameColor())
+        .withTooltip(t -> Common.newTooltip("menu.seasonhud.color.enableSeasonNameColor.tooltip"))
         .create(leftButtonX, MENU_PADDING, BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.color.enableSeasonNameColor.button"), (b, val) -> {
-              Config.setEnableSeasonNameColor(val);
+                Common.translatedText("menu.seasonhud.color.enableSeasonNameColor.button"), (b, val) -> {
+              SeasonHudClient.setEnableSeasonNameColor(val);
               rebuildWidgets();
             });