+++ common/src/main/java/club/iananderson/seasonhud/client/gui/screens/SeasonOptionsScreen.java
@@ -6,21 +6,21 @@ import club.iananderson.seasonhud.client.gui.ShowDay;
 import club.iananderson.seasonhud.client.gui.components.sliders.BasicSlider;
 import club.iananderson.seasonhud.client.gui.components.sliders.HudOffsetSlider;
 import club.iananderson.seasonhud.client.gui.components.sliders.HudScaleSlider;
-import club.iananderson.seasonhud.config.Config;
+import club.iananderson.seasonhud.config.SeasonHudClient;
+import club.iananderson.seasonhud.config.SeasonHudServer;
 import club.iananderson.seasonhud.impl.seasons.CurrentSeason;
 import java.util.Arrays;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.CycleButton;
 import net.minecraft.client.gui.components.EditBox;
-import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import org.jetbrains.annotations.NotNull;
 
 public class SeasonOptionsScreen extends SeasonHudScreen {
-  private static final Component SCREEN_TITLE = Component.translatable("menu.seasonhud.season.title");
+  private static final Component SCREEN_TITLE = Common.translatedText("menu.seasonhud.season.title");
   private Location hudLocation;
   private int xSliderInt;
   private int ySliderInt;
@@ -50,42 +50,45 @@ public class SeasonOptionsScreen extends SeasonHudScreen {
 
   public void loadConfig() {
     drawDefaultHud = Common.drawDefaultHudMenu();
-    hudLocation = Config.getHudLocation();
-    xSliderInt = Config.getHudX();
-    ySliderInt = Config.getHudY();
-    seasonScale = Config.getHudScale();
-    showDay = Config.getShowDay();
-    seasonColor = Config.getEnableSeasonNameColor();
-    showSubSeason = Config.getShowSubSeason();
-    showTropicalSeason = Config.getShowTropicalSeason();
-    needCalendar = Config.getNeedCalendar();
-    enableCalendarDetail = Config.getCalendarDetailMode();
-    dayLength = Config.getDayLength();
+    hudLocation = SeasonHudClient.getHudLocation();
+    xSliderInt = SeasonHudClient.getHudX();
+    ySliderInt = SeasonHudClient.getHudY();
+    seasonScale = SeasonHudClient.getHudScale();
+    showDay = SeasonHudClient.getShowDay();
+    seasonColor = SeasonHudClient.getEnableSeasonNameColor();
+    showSubSeason = SeasonHudClient.getShowSubSeason();
+    showTropicalSeason = SeasonHudClient.getShowTropicalSeason();
+    needCalendar = SeasonHudServer.getNeedCalendar();
+    enableCalendarDetail = SeasonHudServer.getCalendarDetailMode();
+    dayLength = SeasonHudServer.getDayLength();
   }
 
   public void saveConfig() {
     if (drawDefaultHud) {
-      Config.setHudLocation(hudLocationButton.getValue());
-      Config.setHudX(xSlider.getValueInt());
-      Config.setHudY(ySlider.getValueInt());
-      Config.setHudScale(hudScaleSlider.getValueDouble());
+      SeasonHudClient.setHudLocation(hudLocationButton.getValue());
+      SeasonHudClient.setHudX(xSlider.getValueInt());
+      SeasonHudClient.setHudY(ySlider.getValueInt());
+      SeasonHudClient.setHudScale(hudScaleSlider.getValueDouble());
     }
-    Config.setShowDay(showDay);
-    Config.setEnableSeasonNameColor(seasonColor);
+    SeasonHudClient.setShowDay(showDay);
+    SeasonHudClient.setEnableSeasonNameColor(seasonColor);
 
     if (Common.hasSubSeasons()) {
-      Config.setShowSubSeason(showSubSeason);
-      Config.setShowTropicalSeason(showTropicalSeason);
+      SeasonHudClient.setShowSubSeason(showSubSeason);
+      SeasonHudClient.setShowTropicalSeason(showTropicalSeason);
     }
 
-    if (Common.hasCalendarLoaded()) {
-      Config.setNeedCalendar(needCalendar);
-      Config.setCalendarDetailMode(enableCalendarDetail);
+    if (Common.clientSideConfig()) {
+      if (Common.hasCalendarLoaded()) {
+        SeasonHudServer.setCalendarDetailMode(enableCalendarDetail);
+        SeasonHudServer.setNeedCalendar(needCalendar);
+      }
+
+      if (Common.fabricSeasonsLoaded()) {
+        SeasonHudServer.setDayLength(Integer.parseInt(dayLengthBox.getValue()));
+      }
     }
 
-    if (Common.fabricSeasonsLoaded()) {
-      Config.setDayLength(Integer.parseInt(dayLengthBox.getValue()));
-    }
   }
 
   @Override
@@ -111,8 +114,8 @@ public class SeasonOptionsScreen extends SeasonHudScreen {
         .getSeasonHudConfigText(showDay, showSubSeason);
 
     if (drawDefaultHud) {
-      int DEFAULT_X_OFFSET_SCALED = (int) (Config.DEFAULT_X_OFFSET);
-      int DEFAULT_Y_OFFSET_SCALED = (int) (Config.DEFAULT_Y_OFFSET);
+      int DEFAULT_X_OFFSET_SCALED = SeasonHudClient.DEFAULT_X_OFFSET;
+      int DEFAULT_Y_OFFSET_SCALED = SeasonHudClient.DEFAULT_Y_OFFSET;
       seasonScale = hudScaleSlider.getValueDouble();
       int componentWidth = (int) (this.font.width(seasonCombined) * seasonScale);
       int componentHeight = (int) (this.font.lineHeight * seasonScale);
@@ -154,8 +157,8 @@ public class SeasonOptionsScreen extends SeasonHudScreen {
           break;
 
         case CUSTOM:
-          x = (int) ((xSlider.getValueInt()));
-          y = (int) ((ySlider.getValueInt()));
+          x = (xSlider.getValueInt());
+          y = (ySlider.getValueInt());
           break;
       }
     }
@@ -198,32 +201,32 @@ public class SeasonOptionsScreen extends SeasonHudScreen {
     if (drawDefaultHud) {
       row += 1; // Row 1
       hudLocationButton = CycleButton.builder(Location::getLocationName)
-          .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.season.hudLocation.tooltip")))
+          .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.hudLocation.tooltip"))
           .withValues(Location.values())
           .withInitialValue(hudLocation)
           .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                  Component.translatable("menu.seasonhud.season.hudLocation.button"),
+                  Common.translatedText("menu.seasonhud.season.hudLocation.button"),
                   (b, val) -> this.hudLocation = val);
 
-      hudScaleSlider = HudScaleSlider.builder(Component.translatable("menu.seasonhud.season.scale.slider"))
-          .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.season.scale.tooltip")))
-          .withValueRange(Config.HUD_SCALE_MIN, Config.HUD_SCALE_MAX)
+      hudScaleSlider = HudScaleSlider.builder(Common.translatedText("menu.seasonhud.season.scale.slider"))
+          .withTooltip(Common.newTooltip("menu.seasonhud.season.scale.tooltip"))
+          .withValueRange(SeasonHudClient.HUD_SCALE_MIN, SeasonHudClient.HUD_SCALE_MAX)
           .withInitialValue(seasonScale)
-          .withDefaultValue(Config.DEFAULT_SCALE).withStepSize(0.5).withPrecision(1)
+          .withDefaultValue(SeasonHudClient.DEFAULT_SCALE).withStepSize(0.5).withPrecision(1)
           .withBounds(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT)
           .build();
 
       row += 1; // Row 2
-      xSlider = HudOffsetSlider.builder(Component.translatable("menu.seasonhud.season.xOffset.slider"))
-          .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.season.xOffset.tooltip")))
-          .withValues(0, this.maxWidth(seasonCombined), xSliderInt, Config.DEFAULT_X_OFFSET)
+      xSlider = HudOffsetSlider.builder(Common.translatedText("menu.seasonhud.season.xOffset.slider"))
+          .withTooltip(Common.newTooltip("menu.seasonhud.season.xOffset.tooltip"))
+          .withValues(0, this.maxWidth(seasonCombined), xSliderInt, SeasonHudClient.DEFAULT_X_OFFSET)
           .withBounds(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH / 2 - BasicSlider.SLIDER_PADDING,
                       BUTTON_HEIGHT)
           .build();
 
-      ySlider = HudOffsetSlider.builder(Component.translatable("menu.seasonhud.season.yOffset.slider"))
-          .withTooltip(Tooltip.create(Component.translatable("menu.seasonhud.season.yOffset.tooltip")))
-          .withValues(0, this.maxHeight(), ySliderInt, Config.DEFAULT_Y_OFFSET)
+      ySlider = HudOffsetSlider.builder(Common.translatedText("menu.seasonhud.season.yOffset.slider"))
+          .withTooltip(Common.newTooltip("menu.seasonhud.season.yOffset.tooltip"))
+          .withValues(0, this.maxHeight(), ySliderInt, SeasonHudClient.DEFAULT_Y_OFFSET)
           .withBounds(rightButtonX + BUTTON_WIDTH / 2 + BasicSlider.SLIDER_PADDING, (buttonStartY + (row * yOffset)),
                       BUTTON_WIDTH / 2 - BasicSlider.SLIDER_PADDING, BUTTON_HEIGHT)
           .build();
@@ -233,32 +236,31 @@ public class SeasonOptionsScreen extends SeasonHudScreen {
 
     row += 1; // Row 3 (enableMinimapIntegration -> Row 1)
     CycleButton<ShowDay> showDayButton = CycleButton.builder(ShowDay::getDayDisplayName)
-        .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.season.showDay.tooltip")))
+        .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.showDay.tooltip"))
         .withValues(ShowDay.getValues())
         .withInitialValue(showDay)
         .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.season.showDay.button"), (b, val) -> this.showDay = val);
+                Common.translatedText("menu.seasonhud.season.showDay.button"), (b, val) -> this.showDay = val);
 
     CycleButton<Boolean> seasonColorButton = CycleButton.onOffBuilder(seasonColor)
-        .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.color.enableSeasonNameColor.tooltip")))
+        .withTooltip(t -> Common.newTooltip("menu.seasonhud.color.enableSeasonNameColor.tooltip"))
         .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                Component.translatable("menu.seasonhud.color.enableSeasonNameColor.button"),
+                Common.translatedText("menu.seasonhud.color.enableSeasonNameColor.button"),
                 (b, val) -> this.seasonColor = val);
-
     widgets.addAll(Arrays.asList(showDayButton, seasonColorButton));
 
     if (Common.hasSubSeasons()) {
       row += 1; // Row 4 (enableMinimapIntegration -> Row 2)
       CycleButton<Boolean> showSubSeasonButton = CycleButton.onOffBuilder(showSubSeason)
-          .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.season.showSubSeason.tooltip")))
+          .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.showSubSeason.tooltip"))
           .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                  Component.translatable("menu.seasonhud.season.showSubSeason.button"),
+                  Common.translatedText("menu.seasonhud.season.showSubSeason.button"),
                   (b, val) -> this.showSubSeason = val);
 
       CycleButton<Boolean> showTropicalSeasonButton = CycleButton.onOffBuilder(showTropicalSeason)
-          .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.season.showTropicalSeason.tooltip")))
+          .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.showTropicalSeason.tooltip"))
           .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                  Component.translatable("menu.seasonhud.season.showTropicalSeason.button"),
+                  Common.translatedText("menu.seasonhud.season.showTropicalSeason.button"),
                   (b, val) -> this.showTropicalSeason = val);
 
       widgets.addAll(Arrays.asList(showSubSeasonButton, showTropicalSeasonButton));
@@ -288,23 +290,35 @@ public class SeasonOptionsScreen extends SeasonHudScreen {
         }
       });
       dayLengthBox.setHint(Component.literal("" + dayLength).withStyle(ChatFormatting.DARK_GRAY));
+      dayLengthBox.setEditable(Common.clientSideConfig());
+
       widgets.add(dayLengthBox);
     }
 
     if (Common.hasCalendarLoaded()) {
       row += 1; //Row 5 ((enableMinimapIntegration -> Row 3)
       CycleButton<Boolean> needCalendarButton = CycleButton.onOffBuilder(needCalendar)
-          .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.main.needCalendar.tooltip")))
+          .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.needCalendar.tooltip"))
           .create(leftButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                  Component.translatable("menu.seasonhud.main.needCalendar.button"),
+                  Common.translatedText("menu.seasonhud.season.needCalendar.button"),
                   (b, val) -> this.needCalendar = val);
 
+      needCalendarButton.active = Common.clientSideConfig();
+
       CycleButton<Boolean> calendarDetailModeButton = CycleButton.onOffBuilder(enableCalendarDetail)
-          .withTooltip(t -> Tooltip.create(Component.translatable("menu.seasonhud.main.calendarDetail.tooltip")))
+          .withTooltip(t -> Common.newTooltip("menu.seasonhud.season.calendarDetail.tooltip"))
           .create(rightButtonX, (buttonStartY + (row * yOffset)), BUTTON_WIDTH, BUTTON_HEIGHT,
-                  Component.translatable("menu.seasonhud.main.calendarDetail.button"), (b, val) -> {
+                  Common.translatedText("menu.seasonhud.season.calendarDetail.button"), (b, val) -> {
                 this.enableCalendarDetail = val;
               });
+
+      if (!Common.clientSideConfig()) {
+        needCalendarButton.active = false;
+        needCalendarButton.setTooltip(Common.newTooltip("menu.seasonhud.season.serverSide.tooltip"));
+
+        calendarDetailModeButton.active = false;
+        calendarDetailModeButton.setTooltip(Common.newTooltip("menu.seasonhud.season.serverSide.tooltip"));
+      }
 
       widgets.addAll(Arrays.asList(needCalendarButton, calendarDetailModeButton));
     }