+++ common/src/main/java/club/iananderson/seasonhud/config/SeasonHudClient.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 import net.minecraft.client.Minecraft;
 import net.minecraftforge.common.ForgeConfigSpec;
 
-public class Config {
+public class SeasonHudClient {
   public static final ForgeConfigSpec GENERAL_SPEC;
   public static final int DEFAULT_SPRING_COLOR = 16753595;
   public static final int DEFAULT_SUMMER_COLOR = 16705834;
@@ -32,12 +32,9 @@ public class Config {
   private static ForgeConfigSpec.ConfigValue<Integer> winterColor;
   private static ForgeConfigSpec.ConfigValue<Integer> dryColor;
   private static ForgeConfigSpec.ConfigValue<Integer> wetColor;
-  private static ForgeConfigSpec.BooleanValue needCalendar;
-  private static ForgeConfigSpec.BooleanValue calanderDetailMode;
   private static ForgeConfigSpec.BooleanValue showTropicalSeason;
   private static ForgeConfigSpec.BooleanValue showSubSeason;
   private static ForgeConfigSpec.ConfigValue<ShowDay> showDay;
-  private static ForgeConfigSpec.ConfigValue<Integer> dayLength;
   private static ForgeConfigSpec.BooleanValue enableMinimapIntegration;
   private static ForgeConfigSpec.BooleanValue showDefaultWhenMinimapHidden;
   private static ForgeConfigSpec.BooleanValue journeyMapAboveMap;
@@ -49,7 +46,7 @@ public class Config {
     GENERAL_SPEC = builder.build();
   }
 
-  private Config() {
+  private SeasonHudClient() {
   }
 
   private static void setupConfig(ForgeConfigSpec.Builder builder) {
@@ -103,14 +100,6 @@ public class Config {
     builder.pop();
 
     builder.push("Season");
-    needCalendar = builder.comment(
-        "Require the calendar item to be in the players inventory to show the HUD?\n" + "(true/false)\n"
-            + "Default is false.").define("need_calendar", false);
-
-    calanderDetailMode = builder.comment(
-            "Having the calendar item shows the detailed version of the HUD" + "Default is false.")
-        .define("calendar_detail", false);
-
     showTropicalSeason = builder.comment("Show the Tropical seasons (Wet/Dry) in Tropical Biomes.\n"
                                              + "Will not change the season behavior in the biomes.\n" + "(true/false)\n"
                                              + "Default is true.").define("enable_show_tropical_season", true);
@@ -131,10 +120,6 @@ public class Config {
           .defineEnum("enable_show_day", ShowDay.SHOW_DAY,
                       Arrays.asList(ShowDay.NONE, ShowDay.SHOW_DAY, ShowDay.SHOW_WITH_TOTAL_DAYS));
     }
-
-    dayLength = builder.comment(
-        "Change if you are using a Minecraft day length other than vanilla value and using Fabric Seasons.\n"
-            + "Default Minecraft day is 24000.").defineInRange("day_length", 24000, 0, 2147483647);
     builder.pop();
 
     builder.push("Minimap");
@@ -161,20 +146,20 @@ public class Config {
 
   //SeasonHUD
   public static boolean getEnableMod() {
-    return Config.enableMod.get();
+    return SeasonHudClient.enableMod.get();
   }
 
   public static void setEnableMod(boolean enable) {
-    Config.enableMod.set(enable);
+    SeasonHudClient.enableMod.set(enable);
   }
 
   //HUD
   public static Location getHudLocation() {
-    return Config.hudLocation.get();
+    return SeasonHudClient.hudLocation.get();
   }
 
   public static void setHudLocation(Location location) {
-    Config.hudLocation.set(location);
+    SeasonHudClient.hudLocation.set(location);
   }
 
   public static int getHudX() {
@@ -187,7 +172,7 @@ public class Config {
   }
 
   public static void setHudX(int x) {
-    Config.hudX.set(x);
+    SeasonHudClient.hudX.set(x);
   }
 
   public static int getHudY() {
@@ -200,7 +185,7 @@ public class Config {
   }
 
   public static void setHudY(int y) {
-    Config.hudY.set(y);
+    SeasonHudClient.hudY.set(y);
   }
 
   public static double getHudScale() {
@@ -213,7 +198,7 @@ public class Config {
   }
 
   public static void setHudScale(double scale) {
-    Config.hudScale.set(scale);
+    SeasonHudClient.hudScale.set(scale);
   }
 
   //Colors
@@ -222,7 +207,7 @@ public class Config {
   }
 
   public static void setEnableSeasonNameColor(boolean enable) {
-    Config.enableSeasonNameColor.set(enable);
+    SeasonHudClient.enableSeasonNameColor.set(enable);
   }
 
   public static int getSpringColor() {
@@ -230,7 +215,7 @@ public class Config {
   }
 
   public static void setSpringColor(int rgbColor) {
-    Config.springColor.set(rgbColor);
+    SeasonHudClient.springColor.set(rgbColor);
   }
 
   public static int getSummerColor() {
@@ -238,7 +223,7 @@ public class Config {
   }
 
   public static void setSummerColor(int rgbColor) {
-    Config.summerColor.set(rgbColor);
+    SeasonHudClient.summerColor.set(rgbColor);
   }
 
   public static int getAutumnColor() {
@@ -246,7 +231,7 @@ public class Config {
   }
 
   public static void setAutumnColor(int rgbColor) {
-    Config.autumnColor.set(rgbColor);
+    SeasonHudClient.autumnColor.set(rgbColor);
   }
 
   public static int getWinterColor() {
@@ -254,7 +239,7 @@ public class Config {
   }
 
   public static void setWinterColor(int rgbColor) {
-    Config.winterColor.set(rgbColor);
+    SeasonHudClient.winterColor.set(rgbColor);
   }
 
   public static int getDryColor() {
@@ -262,7 +247,7 @@ public class Config {
   }
 
   public static void setDryColor(int rgbColor) {
-    Config.dryColor.set(rgbColor);
+    SeasonHudClient.dryColor.set(rgbColor);
   }
 
   public static int getWetColor() {
@@ -270,24 +255,7 @@ public class Config {
   }
 
   public static void setWetColor(int rgbColor) {
-    Config.wetColor.set(rgbColor);
-  }
-
-  //Season
-  public static boolean getNeedCalendar() {
-    return needCalendar.get();
-  }
-
-  public static void setNeedCalendar(boolean enable) {
-    Config.needCalendar.set(enable);
-  }
-
-  public static boolean getCalendarDetailMode() {
-    return calanderDetailMode.get();
-  }
-
-  public static void setCalendarDetailMode(boolean enable) {
-    Config.calanderDetailMode.set(enable);
+    SeasonHudClient.wetColor.set(rgbColor);
   }
 
   public static boolean getShowTropicalSeason() {
@@ -295,7 +263,7 @@ public class Config {
   }
 
   public static void setShowTropicalSeason(boolean enable) {
-    Config.showTropicalSeason.set(enable);
+    SeasonHudClient.showTropicalSeason.set(enable);
   }
 
   public static boolean getShowSubSeason() {
@@ -303,7 +271,7 @@ public class Config {
   }
 
   public static void setShowSubSeason(boolean enable) {
-    Config.showSubSeason.set(enable);
+    SeasonHudClient.showSubSeason.set(enable);
   }
 
   public static ShowDay getShowDay() {
@@ -311,15 +279,7 @@ public class Config {
   }
 
   public static void setShowDay(ShowDay showDay) {
-    Config.showDay.set(showDay);
-  }
-
-  public static int getDayLength() {
-    return dayLength.get();
-  }
-
-  public static void setDayLength(int length) {
-    Config.dayLength.set(length);
+    SeasonHudClient.showDay.set(showDay);
   }
 
   public static boolean getShowDefaultWhenMinimapHidden() {
@@ -327,16 +287,16 @@ public class Config {
   }
 
   public static void setShowDefaultWhenMinimapHidden(boolean enable) {
-    Config.showDefaultWhenMinimapHidden.set(enable);
+    SeasonHudClient.showDefaultWhenMinimapHidden.set(enable);
   }
 
   //Minimap
   public static boolean getEnableMinimapIntegration() {
-    return Config.enableMinimapIntegration.get();
+    return SeasonHudClient.enableMinimapIntegration.get();
   }
 
   public static void setEnableMinimapIntegration(boolean enable) {
-    Config.enableMinimapIntegration.set(enable);
+    SeasonHudClient.enableMinimapIntegration.set(enable);
   }
 
   //Journeymap
@@ -345,7 +305,7 @@ public class Config {
   }
 
   public static void setJourneyMapAboveMap(boolean enable) {
-    Config.journeyMapAboveMap.set(enable);
+    SeasonHudClient.journeyMapAboveMap.set(enable);
   }
 
   public static boolean getJourneyMapMacOS() {
@@ -353,6 +313,6 @@ public class Config {
   }
 
   public static void setJourneyMapMacOS(boolean enable) {
-    Config.journeyMapMacOS.set(enable);
+    SeasonHudClient.journeyMapMacOS.set(enable);
   }
 }