+++ forge/src/main/java/club/iananderson/seasonhud/forge/SeasonHudForge.java
@@ -1,7 +1,8 @@
 package club.iananderson.seasonhud.forge;
 
 import club.iananderson.seasonhud.Common;
-import club.iananderson.seasonhud.config.Config;
+import club.iananderson.seasonhud.config.SeasonHudClient;
+import club.iananderson.seasonhud.config.SeasonHudServer;
 import club.iananderson.seasonhud.forge.impl.curios.CuriosCompat;
 import club.iananderson.seasonhud.impl.accessories.AccessoriesCompat;
 import club.iananderson.seasonhud.impl.minimaps.CurrentMinimap;
@@ -9,6 +10,7 @@ import club.iananderson.seasonhud.impl.minimaps.SeasonComponent;
 import dev.architectury.utils.Env;
 import dev.architectury.utils.EnvExecutor;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -16,12 +18,16 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(Common.MOD_ID)
 public class SeasonHudForge {
-  public SeasonHudForge(FMLJavaModLoadingContext context) {
-    var modEventBus = context.getModEventBus();
+  public SeasonHudForge() {
+    var modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
     MinecraftForge.EVENT_BUS.register(this);
     Common.init();
 
-    context.registerConfig(ModConfig.Type.CLIENT, Config.GENERAL_SPEC, "seasonhud-client.toml");
+    ModLoadingContext.get()
+        .registerConfig(ModConfig.Type.CLIENT, SeasonHudClient.GENERAL_SPEC, "seasonhud-client.toml");
+
+    ModLoadingContext.get()
+        .registerConfig(ModConfig.Type.SERVER, SeasonHudServer.GENERAL_SPEC, "seasonhud-server.toml");
 
     modEventBus.addListener(SeasonHudForge::onInitialize);
     modEventBus.addListener(SeasonHudForge::ftbChunkSetup);