+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [1.13.9] - 2025-08-25
+
+### Changed
+
+- Added server side config (will be in the "serverconfig" folder of the world)
+- Moved needCalendar and calenderDetailMode to server side config 
+
+### Fixed
+
+- (1.20.1) Fixed crash if using an older version of Forge
+- (1.21.5) Fixed crash when using the 1.21.5 version on 1.21.2-1.21.4
+- Removed the placeholder "Null" season from the color options screen
+
 ## [1.13.8] - 2025-08-07
 
 ### Added